+++ kyrics/src/main/java/com/kyrics/models/SyncedLine.kt
@@ -4,7 +4,7 @@ package com.kyrics.models
  * Core interface for any synchronized line of content.
  * This is the fundamental abstraction that the library works with.
  */
-interface ISyncedLine {
+interface SyncedLine {
     /**
      * Start time in milliseconds
      */